requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
numpy==1.26.4
opencv-python-headless==4.8.1.78
matplotlib==3.8.2
pillow==10.1.0
python-multipart==0.0.6
